Move integration framework context to dedicated module.
The `Context` type is not really related to the `Request` type, and yet
it's currently located in the `Request` module.

This change moves `Context` into its own module, so that contexts and
requests can be kept separate from one another.
